requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
boto3>=1.34.0
python-dotenv>=1.0.0
